Split API tests
