+++ tests/system/test-api-post-search.js
@@ -1,203 +1,7 @@
 const test = require('ava')
 const { apiClient } = require('../helpers/api-client')
-const intersectsGeometry = require('../fixtures/stac/intersectsGeometry.json')
-const noIntersectsGeometry = require('../fixtures/stac/noIntersectsGeometry.json')
-const { refreshIndices } = require('../helpers/es')
 const { randomId } = require('../helpers/utils')
-
-test('/collections', async (t) => {
-  await refreshIndices()
-
-  const response = await apiClient.get('collections')
-
-  t.true(Array.isArray(response.collections))
-  t.true(response.collections.length > 0)
-
-  t.truthy(response.context.returned)
-})
-
-test('GET /collections has a content type of "application/json', async (t) => {
-  const response = await apiClient.get('collections', { resolveBodyOnly: false })
-
-  t.is(response.headers['content-type'], 'application/json; charset=utf-8')
-})
-
-test('/collections/landsat-8-l1', async (t) => {
-  const response = await apiClient.get('collections/landsat-8-l1')
-
-  t.is(response.id, 'landsat-8-l1')
-})
-
-test('GET /collection/landsat-8-l1 has a content type of "application/json', async (t) => {
-  const response = await apiClient.get('collections/landsat-8-l1', { resolveBodyOnly: false })
-
-  t.is(response.headers['content-type'], 'application/json; charset=utf-8')
-})
-
-test('/collections/collection2', async (t) => {
-  const response = await apiClient.get('collections/collection2')
-
-  t.is(response.id, 'collection2')
-})
-
-test('/collections/{collectionId}/items', async (t) => {
-  const response = await apiClient.get('collections/landsat-8-l1/items')
-  t.is(response.type, 'FeatureCollection')
-  t.is(response.features.length, 2)
-  t.is(response.features[0].id, 'LC80100102015082LGN00')
-  t.is(response.features[1].id, 'LC80100102015050LGN00')
-})
-
-test('GET /collections/{collectionId}/items has a content type of "application/geo+json"', async (t) => {
-  const response = await apiClient.get('collections/landsat-8-l1/items', { resolveBodyOnly: false })
-
-  t.is(response.headers['content-type'], 'application/geo+json; charset=utf-8')
-})
-
-test('/collections/{collectionId}/items/{itemId}', async (t) => {
-  const response = await apiClient.get('collections/landsat-8-l1/items/LC80100102015082LGN00')
-  t.is(response.type, 'Feature')
-  t.is(response.id, 'LC80100102015082LGN00')
-})
-
-test('GET /collections/:collectionId/items/:itemId has a content type of "application/geo+json"', async (t) => {
-  const response = await apiClient.get('collections/landsat-8-l1/items/LC80100102015082LGN00', { resolveBodyOnly: false })
-
-  t.is(response.headers['content-type'], 'application/geo+json; charset=utf-8')
-})
-
-test('/collections/{collectionId}/items with bbox 1', async (t) => {
-  const response = await apiClient.post('collections/landsat-8-l1/items', {
-    json: {
-      bbox: [-180, -90, 180, 90]
-    }
-  })
-  t.is(response.type, 'FeatureCollection')
-  t.is(response.features[0].id, 'LC80100102015082LGN00')
-  t.is(response.features[1].id, 'LC80100102015050LGN00')
-})
-
-test('/collections/{collectionId}/items with bbox 2', async (t) => {
-  const response = await apiClient.post('collections/landsat-8-l1/items', {
-    json: {
-      bbox: [-5, -5, 5, 5]
-    }
-  })
-
-  t.is(response.features.length, 0)
-})
-
-test('/collections/{collectionId}/items with bbox and intersects', async (t) => {
-  const response = await apiClient.post('collections/landsat-8-l1/items', {
-    json: {
-      bbox: [-180, -90, 180, 90],
-      intersects: intersectsGeometry
-    }
-  })
-
-  t.truthy(response.context.matched === 2)
-})
-
-test('/collections/{collectionId}/items with time', async (t) => {
-  let response = await apiClient.post('collections/landsat-8-l1/items', {
-    json: {
-      datetime: '2015-02-19T15:06:12.565047+00:00'
-    }
-  })
-  t.is(response.type, 'FeatureCollection')
-  t.is(response.features[0].id, 'LC80100102015050LGN00')
-
-  response = await apiClient.post('collections/landsat-8-l1/items', {
-    json: {
-      datetime: '2015-02-17/2015-02-20'
-    }
-  })
-  t.is(response.type, 'FeatureCollection')
-  t.is(response.features[0].id, 'LC80100102015050LGN00')
-
-  response = await apiClient.post('collections/landsat-8-l1/items', {
-    json: {
-      datetime: '2015-02-19/2015-02-20'
-    }
-  })
-  t.is(
-    response.features[0].id,
-    'LC80100102015050LGN00',
-    'Handles date range without times inclusion issue'
-  )
-})
-
-test('/collections/{collectionId}/items with limit', async (t) => {
-  const response = await apiClient.post('collections/landsat-8-l1/items', {
-    json: {
-      limit: 1
-    }
-  })
-  t.is(response.type, 'FeatureCollection')
-  t.is(response.features.length, 1)
-})
-
-test('/collections/{collectionId}/items with intersects', async (t) => {
-  let response = await apiClient.post('collections/landsat-8-l1/items', {
-    json: {
-      intersects: intersectsGeometry
-    }
-  })
-  t.is(response.type, 'FeatureCollection')
-  t.is(response.features[0].id, 'LC80100102015082LGN00')
-  t.is(response.features[1].id, 'LC80100102015050LGN00')
-
-  response = await apiClient.post('collections/landsat-8-l1/items', {
-    json: {
-      intersects: noIntersectsGeometry
-    }
-  })
-  t.is(response.features.length, 0)
-})
-
-test('/collections/{collectionId}/items with eq query', async (t) => {
-  const response = await apiClient.post('collections/landsat-8-l1/items', {
-    json: {
-      query: {
-        'eo:cloud_cover': {
-          eq: 0.54
-        }
-      }
-    }
-  })
-  t.is(response.features.length, 1)
-  t.is(response.features[0].id, 'LC80100102015050LGN00')
-})
-
-test('/collections/{collectionId}/items with gt lt query', async (t) => {
-  const response = await apiClient.post('collections/landsat-8-l1/items', {
-    json: {
-      query: {
-        'eo:cloud_cover': {
-          gt: 0.5,
-          lt: 0.6
-        }
-      }
-    }
-  })
-  t.is(response.features.length, 1)
-  t.is(response.features[0].id, 'LC80100102015050LGN00')
-})
-
-test('/', async (t) => {
-  const response = await apiClient.get('')
-  t.true(Array.isArray(response.links))
-})
-
-test('GET /search returns an empty list of results for a collection that does not exist', async (t) => {
-  const collectionId = randomId('collection')
-  const searchParams = new URLSearchParams({ collections: [collectionId] })
-
-  const response = await apiClient.get('search', { searchParams })
-
-  t.true(Array.isArray(response.features))
-  t.is(response.features.length, 0)
-})
+const intersectsGeometry = require('../fixtures/stac/intersectsGeometry.json')
 
 test('POST /search returns an empty list of results for a collection that does not exist', async (t) => {
   const response = await apiClient.post('search', {
@@ -248,14 +52,6 @@ test.skip('/search bbox', async (t) => {
     }
   })
   t.is(response.features.length, 0)
-})
-
-test('GET /search has a content type of "application/geo+json; charset=utf-8', async (t) => {
-  const response = await apiClient.get('search', {
-    resolveBodyOnly: false
-  })
-
-  t.is(response.headers['content-type'], 'application/geo+json; charset=utf-8')
 })
 
 test('POST /search has a content type of "application/geo+json; charset=utf-8', async (t) => {
@@ -483,17 +279,6 @@ test('/search collections', async (t) => {
 
   response = await apiClient.post('search', { json: query })
   t.is(response.features.length, 3)
-})
-
-test('GET /conformance returns the expected conformsTo list', async (t) => {
-  const response = await apiClient.get('conformance')
-  t.is(response.conformsTo.length, 13)
-})
-
-test('GET /conformance has a content type of "application/json', async (t) => {
-  const response = await apiClient.get('conformance', { resolveBodyOnly: false })
-
-  t.is(response.headers['content-type'], 'application/json; charset=utf-8')
 })
 
 test.skip('/search preserve geometry in page GET links', async (t) => {
